+++ tests/ElementFinderTest.php
@@ -567,4 +567,24 @@ class ElementFinderTest extends TestCase
       </breakfast_menu>
       ';
     }
+
+
+    final public function testElement()
+    {
+        $page = new ElementFinder('<div><span title="Hello">sdf</span></div>');
+        $first = $page->element('//span')->first();
+        if ($first !== null) {
+            $first->setAttribute('title', 'Changed');
+            self::assertSame('Changed', $first->getAttribute('title'));
+        } else {
+            self::fail('Cant find first title');
+        }
+
+        $second = $page->element('//span')->first();
+        if ($first != null) {
+            self::assertSame('Hello', $second->getAttribute('title'));
+        } else {
+            self::fail('Can find modified title');
+        }
+    }
 }