+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All Notable changes to `ElementFinder` will be documented in this file
 
-## 0.3.0 [Unreleased]
+## 0.3.0 [2018-03-22]
 
 ### Changed  
 - #84 `ElementFinder` become immutable